Split some .util methods to .util.common for same-level imports
